pyproject.toml
[project]
name = "smithers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.123.5",
    "pydantic-ai~=1.21.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff~=0.14.5",
]
